requirements.txt
Django==4.0
djangorestframework==3.13.1
channels==3.0.4
typer==0.4.0
tensorflow==2.6.0
opencv-python==4.5.4.60
numpy==1.19.5